refector: separate MapContainer component from App
