+++ app/src/App.css
@@ -62,11 +62,6 @@ body {
 
 }
 
-.map {
-  width: 100%;
-  height: 100vh;
-}
-
 .date-picker-panel {
   position: absolute;
   height: 3vh;
@@ -168,7 +163,7 @@ body {
   padding: 0.5rem;
 }
 
-.cost{
+.cost {
 	width: 4rem;
   text-align: right;
   padding: 0px 5px;
