+++ app/src/App.tsx
@@ -1,42 +1,21 @@
-import { useEffect, useRef, useState } from 'react'
-import { Avatar, Grid, Flex, Text,Box, Badge, Skeleton, TextArea, Button } from '@radix-ui/themes';
-import { MapContainer, TileLayer, Marker, Popup, useMapEvents } from 'react-leaflet';
+import { useRef, useState } from 'react'
+import { Avatar, Flex, Text,Box, Badge, Skeleton, TextArea, Button } from '@radix-ui/themes';
 import type { LatLngExpression } from 'leaflet';
-import NumberedDivIcon from './NumberedDivIcon';
 import { SortableList } from './components';
 import { ScrollArea } from 'radix-ui';
 import { DayPicker } from 'react-day-picker';
 import { Popover, PopoverTrigger, PopoverContent } from '@radix-ui/react-popover';
 import { format } from 'date-fns';
 import { CalendarDays } from 'lucide-react';
+import type { PlaceInfo, Range } from './PlaceInfo';
+import { MapContainerComponent } from './MapContainer';
+import { getMarkerSrc } from './Marker';
 import '@radix-ui/themes/styles.css';
 import 'react-day-picker/dist/style.css';
 import 'leaflet/dist/leaflet.css';
 import './App.css'
 import './styles.css';
 import './LeafletNumberedMarkers.css'
-
-const zoom_default: number = 15
-
-type PlaceInfo = {
-  day: number;
-  date: string;
-  isStartSection?: boolean;
-  id: number;
-  order?: number;
-  position: LatLngExpression;
-  title: string;
-  detail: string;
-  src: string;
-  address?: string;
-  tags?: string[];
-  cost?: string;
-};
-
-type Range = {
-  from: Date | undefined;
-  to: Date | undefined;
-};
 
 function getBadgeClassName(content: string) {
   if ((content.length)%5 == 0) return 'badge-mango';
@@ -45,31 +24,6 @@ function getBadgeClassName(content: string) {
   if ((content.length)%5 == 3) return 'badge-sunset';
   if ((content.length)%5 == 4) return 'badge-orchid';
   return 'gray'; // default
-}
-
-function ReverseGeocodeMarker({ addNewPlace }: { addNewPlace: (location: PlaceInfo) => void }) {
-  useMapEvents({
-    click: async (e) => {
-      const { lat, lng } = e.latlng;
-      try {
-        const plc: LatLngExpression = [lat, lng];
-        var place = {
-          day: 1,
-          date: new Date().toLocaleDateString(),
-          id: 0,
-          position: plc,
-          title: "",
-          detail: "",
-          src: "https://lh3.googleusercontent.com/gps-cs-s/AC9h4nocKS_n7LfYXrjscSLL_inNyakjdm3YMtFV0NZgHGQA0R2akgrcz3TrI-RrJxVAgVmdg7yAN8ywEZmYZK8UEff7q9cm4tUCWHoqp57KGA5c7Xsq8NnL3yOiLAe02Uz3GXDen0U=w408-h544-k-no",
-        }
-        addNewPlace(place)
-      } catch (error) {
-        console.error('Reverse geocoding failed:', error);
-      }
-    },
-  });
-
-  return null;
 }
 
 function App() {
@@ -150,21 +104,14 @@ function App() {
       src: "https://lh3.googleusercontent.com/gps-cs-s/AC9h4nq5c2znLp8Y7yfXYB-lTB2x2KRnXFcBKV8nmbEEDfPOQsYWYskOE6aWe287ZyI8T-YShGihssVCc_7BqzHto-eSPyfhox2WacVs1AhzNAyc1Rl5_brDKPw-_mEh7a6a60qgMXTa=w408-h544-k-no"
     },
   ];
+  
   const [range, setRange] = useState<Range>({ from: new Date(), to: new Date()});
   const [itinerary, setItinerary] = useState<String[]>(['7/1/2025', '7/2/2025', '7/3/2025', '7/4/2025', '7/5/2025']);
   const [selectedDate, setSelectedDate] = useState<string>('7/5/2025');
   const [items, setItems] = useState<PlaceInfo[]>(markers);
-  const [isAddNewItem, setIsAddNewItem] = useState<boolean|false>(false);
   const itemContainerscrollRef = useRef<HTMLDivElement | null>(null);
   const navSideBarScrollRef = useRef<HTMLDivElement | null>(null);
   const daySectionScrollRefs = useRef<Record<string, HTMLDivElement | null>>({});
-
-  const addNewPlace = (plc: any) => {
-    plc.id = items.length+1
-    plc.order = items.length+1
-    setItems((prev) => [...prev, plc]); 
-    setIsAddNewItem(true);
-  };
 
   const removeItemById = (idToRemove: number) => {
     const filteredItems = items.filter(item => item.id !== idToRemove);
@@ -248,79 +195,13 @@ function App() {
     else if (day%4 == 2) return 'var(--color-tropical-lagoon-whisper)';
     else return 'var(--color-tropical-mango-tin)';
   };
-
-  const getMarkerSrc = (day: number) => {
-    if (day%4 == 0) return "./src/assets/map-pin-rose-sunset.png";
-    else if (day%4 == 1) return "./src/assets/map-pin.png";
-    else if (day%4 == 2) return "./src/assets/map-pin-lagoon-blue.png";
-    else return "./src/assets/map-pin-mango-deep.png";
-  };
   
-  // Auto-scroll to bottom when items change
-  useEffect(() => {
-    if (itemContainerscrollRef.current) {
-      itemContainerscrollRef.current.scrollTo({ top: itemContainerscrollRef.current.scrollHeight, behavior: 'smooth' });
-    }
-  }, [isAddNewItem]);
-
-  useEffect(() => {
-    const fetchAllLocations = async () => {
-      var item = items[items.length-1]; 
-      var updatedItem = null;
-      const [lat, lon] = item.position.toString().split(',');
-      try {
-        const userAgent = { 'User-Agent': 'Journi/1.0',  };
-        const acceptLangAll = { 'Accept-Language': '*' };
-        const acceptLangEn = { 'Accept-Language': 'en' };
-        const resEn = await fetch(
-          `https://nominatim.openstreetmap.org/reverse?lat=${lat}&lon=${lon}&format=json`, 
-          { headers: {...acceptLangEn, ...userAgent} },
-        );
-        const data = await resEn.json();
-        
-        const resLocal = await fetch(
-          `https://nominatim.openstreetmap.org/reverse?lat=${lat}&lon=${lon}&format=json`,
-          { headers: {...acceptLangAll, ...userAgent} },
-        );
-        const dataLocal = await resLocal.json();
-        updatedItem = {
-          ...item,
-          title: (data.name ? data.name : "Untitled") as string,
-          detail: (dataLocal.name === "" ? data.name : dataLocal.name) as string,
-          address: data.display_name as string,
-          tags: [data.type, data.class] as never[],
-        };
-      } catch (err) {
-        console.error('Error fetching location:', err);
-        updatedItem = item; 
-      }
-
-    items[items.length-1] = updatedItem as PlaceInfo;
-    setItems([...items]);
-
-    }
-    fetchAllLocations();
-    setIsAddNewItem(false);
-  }, [isAddNewItem]);
   
   return (
     <>
     <Flex direction="row" gap="1" style={{ height: "100vh", width: "100vw" }}>
        <Box className="map-container">
-        <MapContainer className="map" center={plc1} zoom={zoom_default} scrollWheelZoom={true}>
-          <TileLayer
-            attribution='&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors'
-            url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
-          />
-          <ReverseGeocodeMarker addNewPlace={addNewPlace} />
-          {items.map(({ day, order, position, title, detail }) => (
-            <Marker key={day+","+order} position={position} icon={new NumberedDivIcon({number:order, iconUrl:getMarkerSrc(day)} as any)}>
-              <Popup> 
-                {title} <br/> {detail} <br/>
-              </Popup> 
-            </Marker>
-          ))}
-        </MapContainer>
+        <MapContainerComponent items={items} setItems={setItems} itemContainerscrollRef={itemContainerscrollRef}/>
       </Box>
       <Box width="40vw" height="100vh" className="item-container">
         <ScrollArea.Root className="ScrollAreaRoot">
